Move endstop debug to Endstops
